Move state ownership to child components
Move swiper state + variant sync into MediaSlider, expanded state +
variant reset into MediaGrid. ProductMedia index becomes stateless.
